+++ projects/budget_management/taha-soud/User.js
@@ -1,0 +1,10 @@
+class User {
+  constructor(userName, firstName, lastName, balance = 0) {
+    this.userName = userName;
+    this.firstName = firstName;
+    this.lastName = lastName;
+    this.balance = balance;
+  }
+}
+
+module.exports = User;
