+++ projects/budget_management/taha-soud/UserFactory.js
@@ -1,0 +1,9 @@
+const User = require("./User");
+
+class UserFactory {
+  static createUser(userName, firstName, lastName) {
+    return new User(userName, firstName, lastName);
+  }
+}
+
+module.exports = UserFactory;
